pyproject.toml
[project]
name = "jupyter_to_deploy"
version = "0.1.0"
description = ""
authors = [
    {name = "Theo Lourenção"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.taskipy.tasks]

run = "uvicorn app.main:app --reload"
test = "pytest -v"
kill = "kill -9 $(lsof -t -i:8000)"